remove clean command
